requirements.txt
fastapi==0.115.*
python-dotenv==1.1.*
jinja2==3.1.*
pydantic==2.11.*
